Add files via upload
A classification task performed on the Rice Seed Dataset (Gonen&Jasmine). The prediction is made by a simple Keras Sequential model with 2 hidden layers and the results are compared with the existing traditional machine learning model. The dataset is available at kaggle.
Link for the dataset: https://www.kaggle.com/datasets/seymasa/rice-dataset-gonenjasmine